% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devd.r
\name{devd_log}
\alias{devd_log}
\title{Retrieve current "access log" buffer}
\usage{
devd_log(p, view = TRUE)
}
\arguments{
\item{p}{a [processx::process()] object (the return value from [devd_start])}

\item{view}{view the log contents in the console as well as return it as a character vector?}
}
\value{
(invisibly) a character vector of the current "access log" lines buffer
}
\description{
Returns (invisibly) a character vector of the "access log" lines from `devd` since
the last call to this function.
}
