% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.r
\name{install_devd}
\alias{install_devd}
\title{Install devd server for your platform}
\usage{
install_devd()
}
\description{
This function will attempt to create local directories to store the `devd`
Go binary in and for logs to be kept. It always pulls the latest version of `devd`
and only has support for 64-bit macOS, Windows or Linux environments.
}
